fix: hoist PlacesDailyBreakdown timeline to Places tab (#40)
The Places tab landed in the tabbed-shell refactor (#38) without the
per-day timeline. PlacesDailyBreakdown was still rendered, but buried
two taps deep inside LocationDetailSheet — so the Places tab itself
read as a "list of Copy actions" and the spec acceptance criterion
"Today + Yesterday timeline shows all clusters with start/end times"
(docs/superpowers/specs/2026-05-25-tabbed-app-design.md L241) was
unmet.

Hoist all three pieces from LocationDetailSheet up to PlacesScreen:
- the placesDailySummary compute (clusterLocationsV2 + buildPlacesDailySummary)
- the PlacesDailyBreakdown render, placed below the current-location card
- the handleNamePlaceTap flow (centroid lookup, merge-vs-name branch)
  plus the name/merge modal it triggers

LocationDetailSheet shrinks to its remaining jobs: Known Places CRUD,
retention settings, Export DB, and the three Copy buttons. The
"Copy Daily Summary" handler still needs the per-day summary; it now
computes on-demand on button press (no useMemo needed — single shot).

App.tsx threads db / locationHistory / setKnownPlaces / setError into
PlacesScreen so the naming flow can write back.

Fixes #40
